refactor: developerActInflTrend
