+++ src/anchors/PerceptorBase.tsx
@@ -1,0 +1,3 @@
+export default abstract class PerceptorBase {
+  public abstract run(config: any | null): Promise<void>;
+}
